detailpage

+++ app/src/main/java/com/professionalandroid/apps/baemin_clone_android/src/login/socialregister/models/SocialResponse.kt
@@ -3,7 +3,7 @@ package com.professionalandroid.apps.baemin_clone_android.src.login.socialregist
 data class SocialResponse (
     val result: Result?,
     val isSuccess: Boolean,
-    val code: Long,
+    val code: Int,
     val message: String
 )
